@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.25.0",
+    "@testing-library/user-event": "^14.5.1",
+    "axios": "^1.6.2",
     "buffer": "^6.0.3",
-    "dotenv": "^14.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "5.3",
-    "react-scripts": "5.0.0",
-    "sass": "^1.49.0",
+    "dotenv": "^16.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.69.5",
     "uninstall": "^0.0.0",
     "uniqid": "^5.4.0",
-    "web-vitals": "^2.1.3"
+    "web-vitals": "^3.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
